apis-demo/demo-api-server/requirements.txt
flask-sqlalchemy
flask-restless
psycopg2
requests